requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click @ file:///Users/runner/miniforge3/conda-bld/click_1651215461185/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
dill==0.3.4
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1651670959703/work
Flask-Cors==3.0.10
flatbuffers==2.0
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1651018001541/work
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.1
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpcio_1652355554150/work
gunicorn==20.1.0
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964070496/work
idna==3.3
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1648728616176/work
importlib-resources==5.7.1
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1648147185463/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1651774399431/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1648854440568/work
libclang==14.0.1
Markdown==3.3.7
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1648737587623/work
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1651609723747/work
munkres==1.1.4
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.2
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1649818082882/work
promise==2.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1647961439546/work
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1652392001628/work
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1644357749526/work
six==1.15.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.5.2
tensorflow-macos==2.8.0
tensorflow-metadata==1.8.0
tensorflow-metal==0.4.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1648827305171/work
tqdm==4.64.0
typing_extensions==4.2.0
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1649111980161/work
urllib3==1.26.9
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1651670883478/work
wrapt==1.14.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1649012893348/work
